move as much command line stuff as possible to lowest level
i did this to maintain the separation of module logic and the bottom-up approach wi're going for.
i didnt like how the actual functional connection of units was done at the App project level rather than at the module project level.
this way, it keeps all the relevant logic for a module in one place where it's easy to find it.