use pattern matching to automatically find and run all unit test projects
this does make them run in series instead of in parallel which is slower, so maybe not worth it.
the irony of it is that the time this saves by not having to maintain the task is probably (almost certainly) vanishingly small compared to the extra time waiting for tests to run.
fuckity.